+++ docs-readme.sh
@@ -8,10 +8,10 @@ BASE_URL="https://unstoppabledomains.github.io/resolution"
 CURRENT_VERSION="v$npm_package_version"
 
 echo "# @unstoppabledomains/resolution Documentation\n"
-echo "[Current Version ${CURRENT_VERSION}]($BASE_URL/$CURRENT_VERSION)\n" > $FILE_DESTINATION
+echo "[Current Version ${CURRENT_VERSION}]($BASE_URL/$CURRENT_VERSION/globals.html)\n" > $FILE_DESTINATION
 echo "## Older Versions\n" >> $FILE_DESTINATION
 for dir in $VERSIONS; do
   if [ "$dir" != "$CURRENT_VERSION" ]; then
-    echo "* [$dir]($BASE_URL/$dir)" >> $FILE_DESTINATION
+    echo "* [$dir]($BASE_URL/$dir/globals.html)" >> $FILE_DESTINATION
   fi
 done
