+++ tsconfig.json
@@ -20,8 +20,10 @@
     "module": "commonjs",
     "excludePrivate": true,
     "excludeNotExported": true,
+    "name": "Unstoppabledomains resolution library",
     "inputFiles": [
       "./src/Resolution.ts",
+      "./src/index.ts",
       "./src/publicTypes.ts",
       "./src/errors/configurationError.ts",
       "./src/errors/resolutionError.ts",
